Maintain server side workouts manifest
Fix user home sync status in single player mode.
Sync workouts before rendering user home.
Changes in launcher pages.
Revert unnecessary changes.
