+++ .gitignore
@@ -5,5 +5,3 @@ __pycache__/
 build/
 dist/
 logs/
-.venv/
-scripts/strava_token.txt
