+++ zwift_offline.py
@@ -65,7 +65,7 @@ import online_sync
 import intervals_workouts
 import trainingpeaks_workouts
 import workout_state
-import local_zwift_workouts
+import workouts_manifest
 
 logging.basicConfig(level=os.environ.get("LOGLEVEL", "INFO"))
 logger = logging.getLogger('zoffline')
@@ -213,6 +213,7 @@ class AnonUser(User, AnonymousUserMixin, db.Model):
     first_name = "z"
     last_name = "offline"
     enable_ghosts = os.path.isfile(ENABLEGHOSTS_FILE)
+    is_admin = False
 
     def is_authenticated(self):
         return True
@@ -882,6 +883,19 @@ def save_player_zfile(player_id, folder, filename, content):
     return row
 
 
+def update_workouts_manifest(player_id, folder, filename, content):
+    workouts_manifest.upsert_manifest_entry(os.path.join(STORAGE_DIR, str(player_id), folder), filename, content)
+    timestamp = int(time.time())
+    row = Zfile.query.filter_by(folder=folder, filename=workouts_manifest.MANIFEST_FILENAME, player_id=player_id).first()
+    if not row:
+        row = Zfile(folder=folder, filename=workouts_manifest.MANIFEST_FILENAME, timestamp=timestamp, player_id=player_id)
+        db.session.add(row)
+    else:
+        row.timestamp = timestamp
+    db.session.commit()
+    return row
+
+
 def remove_player_zfiles_by_prefix(player_id, folder, prefix):
     rows = Zfile.query.filter_by(folder=folder, player_id=player_id)
     removed = False
@@ -949,18 +963,6 @@ def resolve_workout_provider_for_player(player_id):
     return workout_state.resolve_active_provider(load_active_workout_provider(player_id), available_workout_providers_for_player(player_id))
 
 
-def local_zwift_workouts_root():
-    return os.path.expanduser('~/Documents/Zwift/Workouts')
-
-
-def export_local_zwift_workout(player_id, filename, content):
-    return local_zwift_workouts.export_workout(local_zwift_workouts_root(), player_id, filename, content)
-
-
-def remove_local_zwift_workouts_by_prefixes(player_id, prefixes):
-    return local_zwift_workouts.remove_prefixed_workouts(local_zwift_workouts_root(), player_id, prefixes)
-
-
 def managed_workout_prefixes(provider):
     if provider == 'intervals-icu':
         return {'intervals-icu-'}
@@ -974,7 +976,7 @@ def clear_managed_workouts_for_provider(player_id, provider, clear_metadata=True
     for prefix in prefixes:
         remove_player_zfiles_by_prefix(player_id, 'customworkouts', prefix)
     if prefixes:
-        remove_local_zwift_workouts_by_prefixes(player_id, prefixes)
+        workouts_manifest.remove_prefixed_workouts(os.path.join(STORAGE_DIR, str(player_id), 'customworkouts'), prefixes)
     if provider == 'intervals-icu' and clear_metadata:
         clear_intervals_workout_metadata(player_id)
 
@@ -989,17 +991,14 @@ def current_workout_sync_status(player_id, provider=None):
             'provider': provider,
             'metadata': None,
             'server_file_exists': False,
-            'local_status': None,
         }
     filename = metadata.get('filename')
     server_file = os.path.join(STORAGE_DIR, str(player_id), 'customworkouts', filename) if filename else ''
-    local_status = local_zwift_workouts.health_report(local_zwift_workouts_root(), player_id, filename) if filename else None
     return {
         'provider': provider,
         'metadata': metadata,
         'server_file_exists': bool(filename and os.path.exists(server_file)),
         'server_file': server_file,
-        'local_status': local_status,
     }
 
 
@@ -1025,7 +1024,7 @@ def sync_intervals_workout_for_player(player_id):
     def store_workout(filename, content, event):
         clear_managed_workouts_for_provider(player_id, 'intervals-icu', clear_metadata=False)
         stored['zfile'] = save_player_zfile(player_id, 'customworkouts', filename, content)
-        stored['local_export'] = export_local_zwift_workout(player_id, filename, content)
+        update_workouts_manifest(player_id, 'customworkouts', filename, content)
         stored['metadata'] = save_intervals_workout_metadata(player_id, event, filename)
 
     try:
@@ -1035,15 +1034,13 @@ def sync_intervals_workout_for_player(player_id):
             return {
                 **result,
                 'sync_status': current_workout_sync_status(player_id, 'intervals-icu'),
-                'message': 'No Intervals.icu workout found for today. Cleared stale managed exports.',
+                'message': 'No Intervals.icu workout found for today.',
             }
         if result['status'] == 'synced':
             sync_status = current_workout_sync_status(player_id, 'intervals-icu')
             return {
                 **result,
-                'local_export': stored.get('local_export'),
                 'sync_status': sync_status,
-                'message': '%s Local Zwift workout catalog prepared.' % result['message'],
             }
         return result
     except Exception as exc:
@@ -1079,11 +1076,10 @@ def sync_trainingpeaks_workout_for_player(player_id):
 
     activate_workout_provider(player_id, 'trainingpeaks')
     clear_managed_workouts_for_provider(player_id, 'trainingpeaks', clear_metadata=False)
-    exports = []
 
     def store_workout(filename, content, workout):
         save_player_zfile(player_id, 'customworkouts', filename, content)
-        exports.append(export_local_zwift_workout(player_id, filename, content))
+        update_workouts_manifest(player_id, 'customworkouts', filename, content)
 
     try:
         result = trainingpeaks_workouts.sync_exported_workouts(folder, store_workout)
@@ -1091,13 +1087,7 @@ def sync_trainingpeaks_workout_for_player(player_id):
             clear_managed_workouts_for_provider(player_id, 'trainingpeaks', clear_metadata=False)
             return {
                 **result,
-                'message': 'No .zwo workouts were found in the TrainingPeaks bridge folder. Cleared stale managed exports.',
-            }
-        if result['status'] == 'synced':
-            return {
-                **result,
-                'local_exports': exports,
-                'message': '%s Local Zwift workout catalog prepared.' % result['message'],
+                'message': 'No .zwo workouts were found in the TrainingPeaks bridge folder.',
             }
         return result
     except Exception as exc:
@@ -1197,23 +1187,12 @@ def intervals(username):
     if request.method == "POST":
         if request.form['athlete_id'] == "" or request.form['api_key'] == "":
             flash("Intervals.icu credentials can't be empty.")
-            return render_template(
-                "intervals.html",
-                username=current_user.username,
-                sync_status=current_workout_sync_status(current_user.player_id, 'intervals-icu'),
-                active_provider=resolve_workout_provider_for_player(current_user.player_id),
-            )
+            return render_template("intervals.html", username=current_user.username)
         encrypt_credentials(file, (request.form['athlete_id'], request.form['api_key']))
         return redirect(url_for('settings', username=current_user.username))
     cred = decrypt_credentials(file)
-    return render_template(
-        "intervals.html",
-        username=current_user.username,
-        aid=cred[0],
-        akey=cred[1],
-        sync_status=current_workout_sync_status(current_user.player_id, 'intervals-icu'),
-        active_provider=resolve_workout_provider_for_player(current_user.player_id),
-    )
+    return render_template("intervals.html", username=current_user.username, aid=cred[0], akey=cred[1],
+        sync_status=current_workout_sync_status(current_user.player_id, 'intervals-icu'))
 
 
 @app.route("/intervals/<username>/sync", methods=["GET"])
@@ -1232,12 +1211,8 @@ def trainingpeaks(username):
         flash('TrainingPeaks bridge folder saved.')
         return redirect(url_for('trainingpeaks', username=current_user.username))
     bridge_folder = load_trainingpeaks_bridge_folder(current_user.player_id)
-    return render_template(
-        "trainingpeaks.html",
-        username=current_user.username,
-        message=trainingpeaks_workouts.PARTNER_ACCESS_MESSAGE,
-        bridge_folder=bridge_folder,
-    )
+    return render_template("trainingpeaks.html", username=current_user.username,
+        message=trainingpeaks_workouts.PARTNER_ACCESS_MESSAGE, bridge_folder=bridge_folder)
 
 
 @app.route("/trainingpeaks/<username>/sync", methods=["GET"])
@@ -1251,18 +1226,14 @@ def trainingpeaks_sync(username):
 @app.route("/user/<username>/")
 @login_required
 def user_home(username):
-    return render_template(
-        "user_home.html",
-        username=current_user.username,
-        enable_ghosts=bool(current_user.enable_ghosts),
-        climbs=CLIMBS,
-        online=get_online(),
-        is_admin=current_user.is_admin,
-        restarting=restarting,
-        restarting_in_minutes=restarting_in_minutes,
-        active_workout_provider=resolve_workout_provider_for_player(current_user.player_id),
-        workout_sync_status=current_workout_sync_status(current_user.player_id),
-    )
+    provider = resolve_workout_provider_for_player(current_user.player_id)
+    if provider == 'intervals-icu':
+        sync_intervals_workout_for_player(current_user.player_id)
+    elif provider == 'trainingpeaks':
+        sync_trainingpeaks_workout_for_player(current_user.player_id)
+    return render_template("user_home.html", username=current_user.username, enable_ghosts=bool(current_user.enable_ghosts), climbs=CLIMBS,
+        online=get_online(), is_admin=current_user.is_admin, restarting=restarting, restarting_in_minutes=restarting_in_minutes,
+        active_workout_provider=provider, workout_sync_status=current_workout_sync_status(current_user.player_id))
 
 def enqueue_player_update(player_id, wa_bytes):
     if not player_id in player_update_queue:
@@ -4644,8 +4615,7 @@ def launch_zwift():
         if MULTIPLAYER:
             return redirect(url_for('login'))
         else:
-            return render_template("user_home.html", username=current_user.username, enable_ghosts=os.path.exists(ENABLEGHOSTS_FILE), online=get_online(),
-                climbs=CLIMBS, is_admin=False, restarting=restarting, restarting_in_minutes=restarting_in_minutes)
+            return redirect(url_for('user_home', username=current_user.username))
     else:
         if MULTIPLAYER:
             return redirect("http://zwift/?code=zwift_refresh_token%s" % fake_refresh_token_with_session_cookie(request.cookies.get('remember_token')), 302)
@@ -4745,14 +4715,6 @@ def start_zwift():
     selected_climb = request.form['climb']
     if selected_climb != 'CALENDAR':
         climb_override[request.remote_addr] = selected_climb
-    provider = resolve_workout_provider_for_player(current_user.player_id)
-    sync_result = None
-    if provider == 'intervals-icu':
-        sync_result = sync_intervals_workout_for_player(current_user.player_id)
-    elif provider == 'trainingpeaks':
-        sync_result = sync_trainingpeaks_workout_for_player(current_user.player_id)
-    if sync_result and sync_result['status'] in ('synced', 'no_workout', 'error'):
-        flash(sync_result['message'])
     return redirect("/ride", 302)
